move indicator factory to indicators package
